Few main directories are being renamed, day9 directory with two files added
