--- src/day4/jumpingStatements/Continue.java
@@ -1,25 +1,0 @@
-package day4.jumpingStatements;
-
-public class Continue {
-
-    public static void main(String[] args) {
-
-        // Стандартный цикл, выводящий значения от 1 до 10
-        for (int i = 1; i <= 10; i++) {
-            // Условие
-            if (i == 5) {
-                // Если условие будет соблюдено, то данная итерация будет прервана,
-                // однако следующая итерация пройдет в обычном порядке
-                continue;
-            }
-            // Вывод значения переменной в консоль
-            System.out.println(i);
-        }
-
-        // Соответственно в консоль выведутся все цифры кроме 5, так как если значение
-        // переменной станет равным 5 - данная итерация будет прервана и до
-        // вывода переменной в консоль дело не дойдет
-
-        // Весь код после оператора continue в рамках цикла выполнен не будет
-    }
-}
