+++ src/day4_Control_Statements/jumpingStatements/Break.java
@@ -1,0 +1,23 @@
+package day4_Control_Statements.jumpingStatements;
+
+public class Break {
+
+    public static void main(String[] args) {
+
+        // Стандартный цикл
+        for (int i = 1; i <= 10; i++) {
+            // Условие
+            if (i == 5) {
+                // Принудительное окончание цикла при соблюдении условия
+                break;
+            }
+            // Вывод значения переменной в консоль
+            System.out.println(i);
+        }
+
+        // Соответственно 4 раза будет осуществлен вывод в консоль, затем цикл будет
+        // остановлен, так как будет достигнуто условие остановки цикла
+
+        // Весь код после оператора break в рамках цикла выполнен не будет
+    }
+}
