--- src/day8/charAt.java
@@ -1,14 +1,0 @@
-package day8;
-
-public class charAt {
-    public static void main(String[] args) {
-
-// charAt()
-// возвращает конкретный символ строки по его индексу
-
-        String s5 = "Welcome";
-
-        System.out.println("The s5 string index 3 element is: " + s5.charAt(3)); // с
-        System.out.println("The s5 string last index element is: " + s5.charAt(6)); // e
-    }
-}
