--- src/day8/length.java
@@ -1,12 +1,0 @@
-package day8;
-
-public class length {
-    public static void main(String[] args) {
-
-// length()
-// определение длины строки
-
-        String s = "Hello World";
-        System.out.println("s length is " + s.length()); // 11
-    }
-}
