--- src/day8/substring.java
@@ -1,15 +1,0 @@
-package day8;
-
-public class substring {
-    public static void main(String[] args) {
-
-// substring()
-// Извлечь подстроку из основной строки
-
-        String s = "Welcome";
-
-        // Нужно указать начальный индекс и конечный индекс НЕ включительно
-        System.out.println(s.substring(0,3)); // Wel
-        
-    }
-}
