--- src/day8/toUpperCase.java
@@ -1,20 +1,0 @@
-package day8;
-
-public class toUpperCase {
-    public static void main(String[] args) {
-
-        String s = "HelLo wOrLd";
-
-// toUpperCase()
-// преобразует все символы передаваемой строки в верхний регистр
-
-        System.out.println(s.toUpperCase()); // HELLO WORLD
-
-// toLowerCase()
-// преобразует все символы передаваемой строки в верхний регистр
-
-        System.out.println(s.toLowerCase()); // hello world
-
-
-    }
-}
