+++ src/day8_StringMethods/charAt.java
@@ -1,0 +1,14 @@
+package day8_StringMethods;
+
+public class charAt {
+    public static void main(String[] args) {
+
+// charAt()
+// возвращает конкретный символ строки по его индексу
+
+        String s5 = "Welcome";
+
+        System.out.println("The s5 string index 3 element is: " + s5.charAt(3)); // с
+        System.out.println("The s5 string last index element is: " + s5.charAt(6)); // e
+    }
+}
