+++ src/day8_StringMethods/length.java
@@ -1,0 +1,12 @@
+package day8_StringMethods;
+
+public class length {
+    public static void main(String[] args) {
+
+// length()
+// определение длины строки
+
+        String s = "Hello World";
+        System.out.println("s length is " + s.length()); // 11
+    }
+}
