+++ src/day8_StringMethods/substring.java
@@ -1,0 +1,15 @@
+package day8_StringMethods;
+
+public class substring {
+    public static void main(String[] args) {
+
+// substring()
+// Извлечь подстроку из основной строки
+
+        String s = "Welcome";
+
+        // Нужно указать начальный индекс и конечный индекс НЕ включительно
+        System.out.println(s.substring(0,3)); // Wel
+        
+    }
+}
